Исправить генератор каталога выпуска
